Rename lang.types to _support.indexing
