+++ python/shark_turbine/kernel/_support/tracing.py
@@ -7,7 +7,7 @@ import warnings
 
 import torch.fx as fx
 
-from ..lang.types import (
+from .indexing import (
     KernelBuffer,
 )
