+++ python/shark_turbine/kernel/lang/__init__.py
@@ -1,2 +1,9 @@
 from .prims import *
-from .types import *
+
+# Include publics from the _support library.
+from .._support.indexing import (
+    Grid,
+    KernelBuffer,
+    SymbolDef,
+    sym,
+)
